+++ lib/openapi_first/validators/headers_validator.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require_relative './parameters_validator'
+
+module OpenapiFirst
+  module Validators
+    class HeadersValidator < ParametersValidator
+      def source_name
+        :header
+      end
+    end
+  end
+end